nodeipam: split out NodeIPAMConfig into pkg/nodeipamconfig
Split configuration out of NodeIPAM into its own cell so that other
packages can pull in NodeIPAM configuration.

The rationale to position this outside the operator is to avoid
shared packages building dependencies into the operator.
