+++ operator/pkg/nodeipam/cell.go
@@ -8,19 +8,18 @@ import (
 	"log/slog"
 
 	"github.com/cilium/hive/cell"
-	"github.com/spf13/pflag"
 	ctrlRuntime "sigs.k8s.io/controller-runtime"
 
 	"github.com/cilium/cilium/operator/pkg/lbipam"
 	k8sClient "github.com/cilium/cilium/pkg/k8s/client"
+	"github.com/cilium/cilium/pkg/nodeipamconfig"
 )
 
 var Cell = cell.Module(
 	"nodeipam",
 	"Node-IPAM",
 
-	cell.Provide(func(r nodeIpamConfig) NodeIPAMConfig { return r }),
-	cell.Config(nodeIpamConfig{}),
+	nodeipamconfig.Cell,
 	cell.Invoke(registerNodeSvcLBReconciler),
 )
 
@@ -30,28 +29,12 @@ type nodeipamCellParams struct {
 	Logger             *slog.Logger
 	Clientset          k8sClient.Clientset
 	CtrlRuntimeManager ctrlRuntime.Manager
-	Config             nodeIpamConfig
+	Config             nodeipamconfig.NodeIPAMConfig
 	SharedConfig       lbipam.SharedConfig
 }
 
-type nodeIpamConfig struct {
-	EnableNodeIPAM bool
-}
-
-func (r nodeIpamConfig) IsEnabled() bool {
-	return r.EnableNodeIPAM
-}
-
-type NodeIPAMConfig interface {
-	IsEnabled() bool
-}
-
-func (r nodeIpamConfig) Flags(flags *pflag.FlagSet) {
-	flags.Bool("enable-node-ipam", r.EnableNodeIPAM, "Enable Node IPAM")
-}
-
 func registerNodeSvcLBReconciler(params nodeipamCellParams) error {
-	if !params.Clientset.IsEnabled() || !params.Config.EnableNodeIPAM {
+	if !params.Clientset.IsEnabled() || !params.Config.IsEnabled() {
 		return nil
 	}
