+++ pkg/metrics/features/operator/cell.go
@@ -14,9 +14,9 @@ import (
 	ciliumenvoyconfig2 "github.com/cilium/cilium/operator/pkg/ciliumenvoyconfig"
 	"github.com/cilium/cilium/operator/pkg/ingress"
 	"github.com/cilium/cilium/operator/pkg/lbipam"
-	"github.com/cilium/cilium/operator/pkg/nodeipam"
 	k8sversion "github.com/cilium/cilium/pkg/k8s/version"
 	"github.com/cilium/cilium/pkg/metrics"
+	"github.com/cilium/cilium/pkg/nodeipamconfig"
 )
 
 var (
@@ -59,7 +59,7 @@ type featuresParams struct {
 	IngressController ingress.IngressConfig
 	LBIPAM            lbipam.Config
 	LBConfig          ciliumenvoyconfig2.LoadBalancerConfig
-	NodeIPAM          nodeipam.NodeIPAMConfig
+	NodeIPAM          nodeipamconfig.NodeIPAMConfig
 }
 
 func (p featuresParams) IsIngressControllerEnabled() bool {
